refactor: split frontend content panel, implement outline viewer
Split the monolithic WorkspaceView and EditorArea into focused,
self-managed components following the chat/ module pattern:

- Add components/content/ with ContentPanel (tab management, content
  loading, save debounce, file:changed/approval event handling),
  ContentEditor (Monaco wrapper), OutlineViewer (Markdown read-only),
  and TabBar (moved from editor/)
- Rename components/editor/ to components/content/, delete EditorArea
- Rename components/novel/ to components/sidebar/; split SidePanel
  into a routing shell and independent NovelList, ChapterList panels
- Make useApp() return a stable reference via useMemo to avoid
  excessive useEffect re-subscriptions causing dropped events
- ChapterList and ContentPanel each listen to file:changed
  independently to refresh sidebar stats and tab contents
- Content and outline are now separate files accessed via the same
  path-based API as the edit tool (chapters/001.md, outlines/001.md)
- Outline view renders Markdown (read-only), content view uses Monaco
- After diff approval/rejection, auto-navigate to the corresponding
  file tab, opening it if necessary
- Approve/reject now handles outline diffs by deriving the matching
  content tab path and refreshing outlineContent
