+++ frontend/src/components/sidebar/BookCover.tsx
