+++ frontend/src/components/sidebar/SidePanel.tsx
@@ -1,0 +1,58 @@
+import type { novel, chapter } from '@/hooks/useApp'
+import NovelList from './NovelList'
+import ChapterList from './ChapterList'
+
+interface Props {
+  activePanel: string
+  novels: novel.Novel[]
+  novelId: number
+  onSelectNovel: (n: novel.Novel) => void
+  onSelectChapter: (ch: chapter.Chapter) => void
+  onSelectGoink: () => void
+  target: { path: string; title: string } | null
+  showCreate: boolean
+  setShowCreate: (v: boolean) => void
+  title: string
+  setTitle: (v: string) => void
+  description: string
+  setDescription: (v: string) => void
+  onCreateNovel: () => void
+}
+
+export default function SidePanel({
+  activePanel,
+  novels, novelId, onSelectNovel,
+  onSelectChapter, onSelectGoink, target,
+  showCreate, setShowCreate, title, setTitle, description, setDescription,
+  onCreateNovel,
+}: Props) {
+  return (
+    <aside className="w-56 border-r bg-background flex flex-col shrink-0">
+      {activePanel === 'novels' ? (
+        <NovelList
+          novels={novels}
+          novelId={novelId}
+          onSelectNovel={onSelectNovel}
+          showCreate={showCreate}
+          setShowCreate={setShowCreate}
+          title={title}
+          setTitle={setTitle}
+          description={description}
+          setDescription={setDescription}
+          onCreateNovel={onCreateNovel}
+        />
+      ) : activePanel === 'chapters' ? (
+        <ChapterList
+          novelId={novelId}
+          target={target}
+          onSelectChapter={onSelectChapter}
+          onSelectGoink={onSelectGoink}
+        />
+      ) : (
+        <div className="flex-1 flex items-center justify-center">
+          <p className="text-xs text-muted-foreground">即将推出</p>
+        </div>
+      )}
+    </aside>
+  )
+}
